chore(refactor): Rename landing page components and assosiated routes
