+++ src/App.scss
@@ -2,7 +2,6 @@
 @import "~@redhat-cloud-services/frontend-components-notifications/index.css";
 @import "~@redhat-cloud-services/frontend-components/index.css";
 @import "~@redhat-cloud-services/frontend-components-utilities/files/Utilities/_all";
-@import "Components/SmartComponents/CVEs/vulnerabilities.scss";
 
 .page__Vulnerabilities {
     overflow-x: auto !important;
