+++ src/Components/SmartComponents/CVEPage/CVEPage.js
@@ -44,8 +44,8 @@ class CVEPage extends React.Component {
 
         setHeader({
             breadcrumbs: [
-                { title: PATHS.vulnerabilities.title, to: PATHS.vulnerabilities.to },
-                { title: PATHS.vulnerabilitiesCves.title, to: PATHS.vulnerabilitiesCves.to },
+                { title: PATHS.home.title, to: PATHS.home.to },
+                { title: PATHS.cves.title, to: PATHS.cves.to },
                 { title: this.state.cveName, isActive: true }
             ]
         });
