+++ src/Components/SmartComponents/CVEs/CVEsTable.js
@@ -11,9 +11,9 @@ import {
     EmptyCVEList
 } from '../../PresentationalComponents/EmptyStates/EmptyStates';
 
-import { CVETableContext } from './VulnerabilitiesCves';
+import { CVETableContext } from './CVEs';
 
-const VulnerabilitiesTableWithContext = ({ context, header, entity }) => {
+const CVEsTableWithContext = ({ context, header, entity }) => {
 
     const noCves = () => {
         const { cves } = context;
@@ -102,16 +102,16 @@ const VulnerabilitiesTableWithContext = ({ context, header, entity }) => {
 
 };
 
-VulnerabilitiesTableWithContext.propTypes = {
+CVEsTableWithContext.propTypes = {
     context: propTypes.any,
     header: propTypes.array,
     entity: propTypes.object
 };
 
-const VulnerabilitiesTable = props => (
+const CVEsTable = props => (
     <CVETableContext.Consumer>
-        {context => <VulnerabilitiesTableWithContext context={context} {...props} />}
+        {context => <CVEsTableWithContext context={context} {...props} />}
     </CVETableContext.Consumer>
 );
 
-export default VulnerabilitiesTable;
+export default CVEsTable;
