+++ src/Components/SmartComponents/CVEs/CVEsTableToolbar.js
@@ -2,7 +2,7 @@ import React from 'react';
 import propTypes from 'prop-types';
 import { injectIntl } from 'react-intl';
 import messages from '../../../Messages';
-import { CVETableContext } from './VulnerabilitiesCves';
+import { CVETableContext } from './CVEs';
 import { fetchCvesIds } from '../../../Store/Actions/Actions';
 import selectAllCheckbox from '../../../Helpers/selectAllCheckboxHelper';
 import { PrimaryToolbar } from '@redhat-cloud-services/frontend-components/components/cjs/PrimaryToolbar';
@@ -22,7 +22,7 @@ import {
     removeFilters
 } from '../../../Helpers/TableToolbarHelper';
 
-const  VulnerabilitiesToolbarWithContext = (props) => {
+const CVEsTableToolbarWithContext = (props) => {
     const handleCveDescription = () => {
         const { context } = props;
         const { cves, methods, expandCveDescription } = context;
@@ -111,20 +111,20 @@ const  VulnerabilitiesToolbarWithContext = (props) => {
     );
 };
 
-VulnerabilitiesToolbarWithContext.propTypes = {
+CVEsTableToolbarWithContext.propTypes = {
     context: propTypes.object,
     intl: propTypes.object
 };
 
-VulnerabilitiesToolbarWithContext.defaultProps = {
+CVEsTableToolbarWithContext.defaultProps = {
     totalNumber: 0,
     apply: () => undefined,
     downloadReport: () => undefined
 };
 
-const VulnerabilitiesToolbar = props => (
+const CVEsTableToolbar = props => (
     <CVETableContext.Consumer>
-        {context => <VulnerabilitiesToolbarWithContext context={context} {...props} />}
+        {context => <CVEsTableToolbarWithContext context={context} {...props} />}
     </CVETableContext.Consumer>
 );
-export default injectIntl(VulnerabilitiesToolbar);
+export default injectIntl(CVEsTableToolbar);
