--- src/Components/SmartComponents/CVEs/vulnerabilities.scss
@@ -1,49 +1,0 @@
-@import "~@redhat-cloud-services/frontend-components-utilities/files/Utilities/_mixins";
-
-.page__Vulnerabilities {
-    @media screen and (max-width: 1100px) {
-        .pf-c-pagination:not(.pf-m-footer) .pf-c-options-menu,
-        .pf-c-pagination:not(.pf-m-footer) .pf-c-pagination__nav {
-            display: none;
-            visibility: hidden;
-        }
-        .pf-c-pagination__total-items {
-            display: flex;
-            visibility: visible;
-        }
-    }
-
-    p.pf-c-empty-state__body {
-        ul {
-            margin-top: 0px;
-            margin-bottom: var(--pf-global--spacer--sm);
-            li {
-                margin-top: 0px;
-            }
-        }
-    }
-
-    .hide-description {
-        @media (max-width: 1500px) and (min-width: 992px) {
-            --pf-hidden-visible--Display: var(
-                --pf-hidden-visible--hidden--Display
-            );
-            --pf-hidden-visible--Visibility: var(
-                --pf-hidden-visible--hidden--Visibility
-            );
-        }
-    }
-
-    .col-width-10 {
-        width: 10%;
-    }
-    .no-wrap {
-        white-space: nowrap;
-    }
-    tr.pf-m-expanded {
-        .pf-c-table__action {
-            display: none;
-        }
-    }
-}
-
