+++ src/Components/SmartComponents/LandingPage/LandingPage.js
@@ -1,0 +1,18 @@
+import React, { Fragment } from 'react';
+import { Main } from '@redhat-cloud-services/frontend-components';
+import VulnerabilityHeader from '../VulnerabilityHeader/VulnerabilityHeader';
+import CVEs from '../CVEs/CVEs';
+
+const LandingPage = () => {
+
+    return (
+        <Fragment>
+            <VulnerabilityHeader showBreadcrumb={false} showDownloadButton />
+            <Main>
+                <CVEs/>
+            </Main>
+        </Fragment>
+    );
+};
+
+export default LandingPage;
