+++ src/Components/SmartComponents/SystemCves/SystemCves.js
@@ -8,7 +8,6 @@ import {
     fetchSystemDetailsIds
 } from '../../../Store/Actions/Actions';
 import propTypes from 'prop-types';
-import '../CVEs/vulnerabilities.scss';
 import messages from '../../../Messages';
 import { withRouter } from 'react-router-dom';
 import SystemCveTable from './SystemCveTable';
