+++ src/Components/SmartComponents/SystemDetail/SystemDetail.js
@@ -26,7 +26,7 @@ class SystemDetail extends React.Component {
         const { entity, isOptOut, optOutSystemHandler, loaded } = this.props;
         this.props.setHeader({
             breadcrumbs: [
-                { title: PATHS.vulnerabilities.title, to: PATHS.vulnerabilities.to },
+                { title: PATHS.home.title, to: PATHS.home.to },
                 { title: PATHS.systems.title, to: PATHS.systems.to },
                 { title: (entity && entity.display_name) || <FormattedMessage {...messages.invalidSystem} />, isActive: true }
             ]
