+++ src/Helpers/constants.js
@@ -20,11 +20,11 @@ export const impactTextList = {
 };
 
 export const PATHS = {
-    vulnerabilities: {
+    home: {
         title: 'Vulnerability',
         to: '/'
     },
-    vulnerabilitiesCves: {
+    cves: {
         title: 'CVEs',
         to: '/cves'
     },
