+++ src/Utilities/Routes.js
@@ -8,8 +8,8 @@ import { PATHS } from '../Helpers/constants';
 const CVEPage = asyncComponent(() => import(/* webpackChunkName: "CVEPage" */ '../Components/SmartComponents/CVEPage/CVEPage'));
 const Systems = asyncComponent(() => import(/* webpackChunkName: "Systems" */ '../Components/SmartComponents/Systems/Systems'));
 
-const VulnerabilitiesCVEs = asyncComponent(() =>
-    import(/* webpackChunkName: "VulnerabilitiesCVEs" */ '../Components/SmartComponents/CVEs/CVEs'));
+const LandingPage = asyncComponent(() =>
+    import(/* webpackChunkName: "LandingPage" */ '../Components/SmartComponents/LandingPage/LandingPage'));
 const InventoryDetail = asyncComponent(() =>
     import(/* webpackChunkName: "InventoryDetail" */ '../Components/SmartComponents/InventoryDetail/InventoryDetail')
 );
@@ -41,11 +41,11 @@ export const Routes = () => {
             <Redirect from={PATHS.cvepagePath.to} to={PATHS.inventoryDetail.to} />
             <InsightsRoute exact path={PATHS.cvePage.to} component={CVEPage} />
             <InsightsRoute exact path={PATHS.inventoryDetail.to}  component={InventoryDetail} />
-            <InsightsRoute exact path={PATHS.vulnerabilitiesCves.to} component={VulnerabilitiesCVEs} />
+            <InsightsRoute exact path={PATHS.cves.to} component={LandingPage} />
             <InsightsRoute exact path={PATHS.systems.to} component={Systems} />
             <InsightsRoute exact path={PATHS.upgrade.to} component={Upgrade} />
             <Route
-                render={() => (some(PATHS, p => p.to === path) ? null : <Redirect to={PATHS.vulnerabilitiesCves.to} />)}
+                render={() => (some(PATHS, p => p.to === path) ? null : <Redirect to={PATHS.cves.to} />)}
             />
         </Switch>
     );
